+++ src/counter.rs
@@ -2,6 +2,7 @@
 // License, v. 2.0. If a copy of the MPL was not distributed with this
 // file, You can obtain one at http://mozilla.org/MPL/2.0/.
 
+#[cfg(feature = "sync")]
 use std::sync::atomic::{AtomicUsize, Ordering};
 
 #[doc(hidden)]
@@ -30,6 +31,7 @@ impl Counter for usize {
     }
 }
 
+#[cfg(feature = "sync")]
 impl Counter for AtomicUsize {
     #[inline(always)]
     fn inc(&mut self) {