+++ src/sync_type.rs
@@ -11,9 +11,11 @@ use crate::pool::PoolInner;
 use crate::stack::Stack;
 
 #[cfg(feature = "sync")]
+use crate::pointer::NonNullAtomicPtr;
+#[cfg(feature = "sync")]
 use crossbeam_queue::ArrayQueue;
 #[cfg(feature = "sync")]
-use std::sync::atomic::{AtomicPtr, AtomicUsize};
+use std::sync::atomic::AtomicUsize;
 
 #[doc(hidden)]
 pub trait PoolSyncType<A>: Sized {
@@ -36,8 +38,8 @@ pub struct PoolUnsync;
 impl<A> PoolSyncType<A> for PoolSync {
     type Counter = AtomicUsize;
     type Stack = ArrayQueue<Self::ElementPointer>;
-    type ElementPointer = AtomicPtr<RefBox<A, Self>>;
-    type PoolPointer = AtomicPtr<PoolInner<A, Self>>;
+    type ElementPointer = NonNullAtomicPtr<RefBox<A, Self>>;
+    type PoolPointer = NonNullAtomicPtr<PoolInner<A, Self>>;
 }
 
 impl<A> PoolSyncType<A> for PoolUnsync {