+++ CHANGELOG.md
@@ -9,7 +9,7 @@ adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ### FIXED
 
--   The unsync `Pool` and `PoolRef` now use `NonNull` instead of raw pointers, so that they can be
+-   The `Pool` and `PoolRef` now use `NonNull` instead of raw pointers, so that they can be
     wrapped in `Option` without growing in size.
 -   Fixed a race condition where the last `PoolRef` referencing a pool might try to drop it before
     returning its allocation to it.